use max of retail rate or 10% over negotiated rate in shipengine.endpts
It looks like requested_comparison_amount is included in response from shipengine, but the feature is not currently active on the api.shipengine.com/v1/rates/estimate route.

For now, it's baked into the code, but shipping amount defaults to max of retail rate or 10% upcharge over negotiated rate.

In future shipping options, we can offer the retail rate or up to that upcharge rate as a convenience fee.